Small updates, little improvements
