+++ .gitignore
@@ -4,3 +4,13 @@
 *.lock
 
 Lucien-Rust/target/
+
+\.idea/
+
+cmake-build-debug/
+
+Lucien/Rust/target/
+
+Lucien/Rust/Rust\.iml
+
+Lucien/server/
